First project commit
A lot of bugs. I'm pretty sure it'll crash on the 5th if not the 3rd activity. I could try to fix them myself but I don't have a suitable access to computer at the time, that's a hard fact. There's a lot to find, fix and add.

I'd be glad to help on anything concerning the code. Enjoy debugging and improving the project!